🐞 fix(installation): handle tipe 'None' jika dokumen tidak ditemukan

+++ app/core/setup.py
@@ -224,15 +224,16 @@ async def unregister_payment_methods():
             pm_obj = await collections.payment_method.find_one(
                 {"pg_id": pg_id, "code": pm.value}
             )
-            pm_id = str(pm_obj["_id"])
-            print(f"  * Unregistering payment method: {pm.value} ({channel_code})")
-            pc_filter = {"pm_id": pm_id, "code": channel_code}
-            result: DeleteResult = await collections.payment_channel.delete_one(
-                pc_filter
-            )
-            msg = (
-                "   * Payment channel deleted"
-                if result.deleted_count > 0
-                else "   ! Payment channel already deleted"
-            )
-            print(msg)
+            if pm_obj:
+                pm_id = str(pm_obj["_id"])
+                print(f"  * Unregistering payment method: {pm.value} ({channel_code})")
+                pc_filter = {"pm_id": pm_id, "code": channel_code}
+                result: DeleteResult = await collections.payment_channel.delete_one(
+                    pc_filter
+                )
+                msg = (
+                    "   * Payment channel deleted"
+                    if result.deleted_count > 0
+                    else "   ! Payment channel already deleted"
+                )
+                print(msg)
